Update README.md
This project is only a humble venture to satisfy the needs in a hospital system. 
Several user friendly coding have also adopted. This package shall prove to be a powerful 
package in satisfying all the requirements of the Hospital Management System.
 The objective of software planning is to provide a frame work that enables the 
hospital staff to make reasonable time consuming ways made within a limited time frame at 
the beginning of the software project and should be updated regularly as the project 
progresses.